F12 §5: the PKCS#11 layer runs on Linux, and the PIN dialog is a GTK window
D-349. internal/keysource/pkcs11 could not reach a token on this platform —
openModule refused — so the dialog it exists for had nothing to talk to. The
Linux layer is cgo against the platform's own pkcs11.h rather than a second
hand-written layout, because F11 measured three of four candidate layouts
returning CKR_OK with a zero length. RTLD_LOCAL so a vendor module's OpenSSL
stays its own; RTLD_NOW so a missing symbol fails at the load with a reason.

The neutral logic is shared rather than copied. ckULong is a type alias —
uint32 on Windows, uint64 here — so sign.go, source.go and login.go name it
once and neither platform's code had to be rewritten for the other. Copying
privateKeyFor and Open would have been this session's own D-346 a third time,
in security code.

Run against SoftHSM: a 256-byte signature verified against the certificate's
own public key, one wrong PIN costing exactly one attempt, and clause 1's
protected-path flag clear on a third module. It found something immediately:
MaxPINLength refused a token declaring ulMaxPinLen 255, which is a conforming
software token rather than an absurd one. The declared maximum is clamped now;
a token whose minimum exceeds what this layer will hold is still refused.

And pin_test.go refused the first refactor, correctly: passing the PIN to a
per-platform callLogin makes it travel to a second function, which clause 2
forbids. login.go now holds only the arithmetic and never sees a PIN; the
twenty lines that hold the characters are per platform because the call at the
end of them is.

D-350. The dialog is native, and the measurement says clause 2 needs no third
exception: GtkPasswordEntry's buffer is mlocked, holds exactly one copy of what
was typed, and is emptied by overwriting through the widget's own interface —
which is exception 1 in GTK's spelling, and stronger, since the Windows edit
control's buffer is pageable.

Two instruments had to be fixed first. The memory scan found two copies before
anything was set — its own needle and its own scan buffer — and an instrument
that finds itself cannot report what is new. The undo check said "restored
nothing" and its control said the same on a plain entry, which meant it was
measuring a typo; gtk_editable_insert_text does not feed GTK's undo history,
and insert-at-cursor does.

The PIN never becomes a Go string: three C helpers give a length, a copy into
the caller's buffer, and an overwrite. The AST guard names the binding's Text()
accessor by name, because that is the obvious thing to reach for and it would
compile.

The GTK boundary is three packages now. pinscreen opens a window, so it links
what a window links — and the guard's rule is restated as an argument rather
than a count.

Nobody has typed a PIN into this window yet. Recorded as unwatched.

+++ internal/keysource/pkcs11/certificate.go
@@ -69,10 +69,16 @@ var ErrCertificateNotFound = errors.New("pkcs11: no certificate with that thumbp
 // Thumbprint is what collapses them; the rest is what says which sighting a
 // row came from (F11 §4).
 type CertificateInfo struct {
-	Thumbprint   string
-	DER          []byte
-	Label        string
-	SlotID       uint32
+	Thumbprint string
+	DER        []byte
+	Label      string
+	// SlotID is a CK_ULONG, which is four bytes on Windows and eight on
+	// Linux (D-349). It is declared at the wider of the two here rather
+	// than per platform, because a field on an exported struct that
+	// changes width between builds is a field every caller has to know
+	// the platform of — and widening is the only direction that cannot
+	// lose one.
+	SlotID       uint64
 	TokenLabel   string
 	TokenSerial  string
 	ModulePath   string
